creative-lab/story-generator/requirements.txt
fastapi
uvicorn
openai
python-dotenv  # (optional, for safely loading API keys)
